+++ tools/src/check-packages/checkDependenciesAsync.ts
@@ -33,11 +33,7 @@ type SourceFileImportRef = {
   isTypeOnly?: boolean;
 };
 
-// We are incrementally rolling this out, the sdk packages in this list are expected to be invalid
-const IGNORED_PACKAGES = [
-  '@expo/html-elements', // package: react, react-native, react-native-web
-  'expo-gl', // package: react-dom, react-native-reanimated
-];
+const IGNORED_PACKAGES = [];
 
 const SPECIAL_DEPENDENCIES: Record<string, Record<string, IgnoreKind | void> | void> = {
   'expo-dev-menu': {
@@ -45,15 +41,6 @@ const SPECIAL_DEPENDENCIES: Record<string, Record<string, IgnoreKind | void> | v
   },
   'expo-modules-test-core': {
     typescript: 'ignore-dev', // TODO: Should probably be a peer dep
-  },
-
-  'expo-asset': {
-    '@expo/config-plugins/build/utils/warnings.js': 'ignore-dev', // TODO: Remove
-  },
-
-  'expo-font': {
-    '@expo/config-plugins/build/android/codeMod': 'ignore-dev', // TODO: Remove
-    '@expo/config-plugins/build/utils/generateCode': 'ignore-dev', // TODO: Remove
   },
 
   '@expo/cli': {